sure pre-commit have installed

+++ test/init.sh
@@ -1,5 +1,11 @@
 #! /bin/bash
 
+which shellcheck &> /dev/null
+if [[ $? != 0 ]]; then
+    echo "are you sure you have installed shellcheck?"
+    exit 1
+fi
+
 cat << EOS > .pre-commit-config.yaml
 -   repo: $(pwd)
     sha: $(git rev-parse HEAD)
